+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.5.5 (tbd)
 
 - Fixed `piral-ng/extend-webpack` for MJS files using not fully specified references
+- Added `piral-ng/standalone` to support pure modern Angular standalone (#690)
 
 ## 1.5.4 (April 23, 2024)